fix(mlx): make float32 precision contract explicit and harden tests
Deep-dive into the cross-machine non-determinism flagged in review: the
MLX backend evaluates the logp/gradient on the Metal GPU, where float64
is unsupported, so ``mx.array(_x)`` silently downcasts nutpie's float64
positions to float32. Measured relative error is ~1e-7 (float32 eps), and
because NUTS is a chaotic integrator that error accumulates across the
leapfrog steps and varies with GPU/Metal/kernel fusion -- which is why the
array_compare reference test could never match between machines.

* Cast inputs explicitly to ``mx.float32`` in the logp and expand closures
  so the precision contract is intentional and cannot silently change with
  the MLX version or a global default-dtype override, and document the
  root cause at the conversion site.

* Replace the bare smoke test with a machine-independent statistical
  correctness check: sample HalfNormal and assert the posterior recovers
  the analytic moments within ~4x the Monte Carlo error.

* xfail ``test_pymc_model_store_extra[mlx-pytensor]``: PyTensor's MLX
  linker Split dispatch passes a numpy array to ``mx.split`` (read as an
  equal-split count) and breaks on the uneven splits in the
  ZeroSumNormal/Dirichlet gradient graphs. mlx/mlx is unaffected.

Made-with: Cursor

+++ python/nutpie/compile_pymc.py
@@ -605,7 +605,15 @@ def _compile_pymc_model_mlx(
 
     def make_logp_func():
         def logp(_x, **shared):
-            _x_mlx = mx.array(_x)
+            # nutpie operates in float64, but MLX evaluates on the Metal GPU
+            # where float64 is unsupported, so the logp and its gradient are
+            # necessarily computed in float32. We cast explicitly rather than
+            # relying on ``mx.array``'s implicit float64->float32 downcast, so
+            # the precision contract cannot silently change with the MLX
+            # version or a global default-dtype override. This single-precision
+            # evaluation is the root cause of MLX sampling not being
+            # bit-reproducible across machines; see test_deterministic_sampling_mlx.
+            _x_mlx = mx.array(_x, dtype=mx.float32)
             logp, grad = logp_fn(_x_mlx, *[shared[name] for name in logp_shared_names])
             return float(logp), np.asarray(grad, dtype="float64", order="C")
 
@@ -618,7 +626,9 @@ def _compile_pymc_model_mlx(
     def make_expand_func(seed1, seed2, chain):
         # TODO handle seeds
         def expand(_x, **shared):
-            _x_mlx = mx.array(_x)
+            # Match the logp closure: cast explicitly to float32 (MLX has no
+            # float64 on the GPU) instead of relying on the implicit downcast.
+            _x_mlx = mx.array(_x, dtype=mx.float32)
             values = expand_fn(_x_mlx, *[shared[name] for name in expand_shared_names])
             return {
                 name: np.asarray(val, order="C", dtype=dtype).reshape(shape)
